Move createPortal into spending bucket element
The useEffect, depending only on the map, prevents the re-render which
motivated moving the div ref up to the map container.
