+++ src/app/src/components/AnimatedArcsAndMap/AnimatedTotalSpendingBucket.tsx
@@ -1,14 +1,34 @@
 import { useEffect, useState } from 'react';
+import { createPortal } from 'react-dom';
 import { Progress, Tag, TagLabel } from '@chakra-ui/react';
+import { useMap } from 'react-leaflet';
+import L from 'leaflet';
 
 import { SpendingByGeographyAtMonth } from '../../types/api';
-import { TOTAL_BIL_AMOUNT } from '../../constants';
+import { DC_CENTER, TOTAL_BIL_AMOUNT } from '../../constants';
 
 export default function AnimatedTotalSpendingBucket({
     spendingAtTimeByState,
 }: {
     spendingAtTimeByState: SpendingByGeographyAtMonth | undefined;
 }) {
+    const map = useMap();
+    const [div, setDiv] = useState<HTMLDivElement | undefined>();
+
+    useEffect(() => {
+        const div = document.createElement('div');
+        L.popup({ className: 'portal-container', closeButton: false })
+            .setLatLng(DC_CENTER)
+            .setContent(div)
+            .openOn(map);
+
+        setDiv(div);
+
+        return () => {
+            setDiv(undefined);
+        };
+    }, [map]);
+
     const [totalSpendingAtTime, setTotalSpendingAtTime] = useState(0);
 
     useEffect(() => {
@@ -25,7 +45,11 @@ export default function AnimatedTotalSpendingBucket({
 
     const amountLeft = TOTAL_BIL_AMOUNT - totalSpendingAtTime;
 
-    return (
+    if (!div) {
+        return null;
+    }
+
+    return createPortal(
         <Tag background='none' paddingLeft={0}>
             <Progress
                 value={amountLeft}
@@ -46,7 +70,8 @@ export default function AnimatedTotalSpendingBucket({
             >
                 ${divideByBillion(amountLeft)}B
             </TagLabel>
-        </Tag>
+        </Tag>,
+        div
     );
 }
